move all tests to one file to prevent singleton problems
